+++ app/Plugin/Cli.php
@@ -10,11 +10,15 @@ namespace ExternalFilesInMediaLibrary\Plugin;
 // prevent direct access.
 defined( 'ABSPATH' ) || exit;
 
+use easyDirectoryListingForWordPress\Taxonomy;
+use ExternalFilesInMediaLibrary\ExternalFiles\Export;
 use ExternalFilesInMediaLibrary\ExternalFiles\Extensions\Availability;
 use ExternalFilesInMediaLibrary\ExternalFiles\Extensions\Queue;
 use ExternalFilesInMediaLibrary\ExternalFiles\File;
 use ExternalFilesInMediaLibrary\ExternalFiles\Files;
 use ExternalFilesInMediaLibrary\ExternalFiles\Import;
+use ExternalFilesInMediaLibrary\ExternalFiles\Synchronization;
+use WP_Term_Query;
 
 /**
  * Handle external files via WP CLI.
@@ -47,7 +51,7 @@ class Cli {
 	 * [--use_specific_date=<value>]
 	 * : Use specific date for each file.
 	 *
-	 * @param array<string,string> $urls Array with URL, which might be given as parameter on CLI-command.
+	 * @param array<int,string>    $urls Array with URL, which might be given as parameter on CLI-command.
 	 * @param array<string,string> $arguments List of parameter to use for the given URLs.
 	 *
 	 * @return void
@@ -135,7 +139,7 @@ class Cli {
 	 * [<URLs>]
 	 * : List of URLs to delete from in media library. If nothing is given all external files are deleted.
 	 *
-	 * @param array<string> $urls List of URLs.
+	 * @param array<int,string> $urls List of URLs.
 	 *
 	 * @return void
 	 */
@@ -208,15 +212,18 @@ class Cli {
 
 		// log what we have done.
 		$logs->create( __( 'Log has been cleared via cli.', 'external-files-in-media-library' ), '', 'success', 2 );
+
+		// return ok-message.
+		\WP_CLI::success( __( 'Log has been cleared.', 'external-files-in-media-library' ) );
 	}
 
 	/**
-	 * Check all URLs in the media library.
+	 * Check the availability of URLs in the media library.
 	 *
 	 * [<URLs>]
 	 * : List of URLs to check. They must exist in media library.
 	 *
-	 * @param array<string> $urls List of URLs.
+	 * @param array<int,string> $urls List of URLs.
 	 *
 	 * @return void
 	 */
@@ -257,7 +264,7 @@ class Cli {
 		}
 
 		// return ok-message.
-		\WP_CLI::success( __( 'URL-check has been run.', 'external-files-in-media-library' ) );
+		\WP_CLI::success( __( 'URL-check has been run. Results could be checked under Settings > External Files in Media Library > Logs.', 'external-files-in-media-library' ) );
 	}
 
 	/**
@@ -295,7 +302,7 @@ class Cli {
 	 * [<URLs>]
 	 * : List of URLs to switch.
 	 *
-	 * @param array<string> $urls List of URLs.
+	 * @param array<int,string> $urls List of URLs.
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
@@ -332,7 +339,7 @@ class Cli {
 		}
 
 		// return ok-message.
-		\WP_CLI::success( __( 'Files have been switches to external hosting.', 'external-files-in-media-library' ) );
+		\WP_CLI::success( _n( 'File has been switched to external hosting.', 'Files have been switched to external hosting.', count( $files ), 'external-files-in-media-library' ) );
 	}
 
 	/**
@@ -344,7 +351,7 @@ class Cli {
 	 * [<URLs>]
 	 * : List of URLs to switch.
 	 *
-	 * @param array<string> $urls List of URLs.
+	 * @param array<int,string> $urls List of URLs.
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
@@ -381,7 +388,7 @@ class Cli {
 		}
 
 		// return ok-message.
-		\WP_CLI::success( __( 'Files have been switches to local hosting.', 'external-files-in-media-library' ) );
+		\WP_CLI::success( _n( 'File has been switched to local hosting.', 'Files have been switched to local hosting.', count( $files ), 'external-files-in-media-library' ) );
 	}
 
 	/**
@@ -398,7 +405,7 @@ class Cli {
 	}
 
 	/**
-	 * Clear the queue (delete every entry).
+	 * Clear the queue (will delete every entry in the queue).
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
@@ -417,9 +424,313 @@ class Cli {
 	 * @noinspection PhpUnused
 	 */
 	public function cleanup_queue(): void {
-		Queue::get_instance()->clear();
+		Queue::get_instance()->clear_error_entries();
 
-		// return ok-message.
+		// show ok message.
 		\WP_CLI::success( __( 'Error-URLs has been removed from the queue.', 'external-files-in-media-library' ) );
+	}
+
+	/**
+	 * Add an external source.
+	 *
+	 * <Name>
+	 * : The name to use to this entry.
+	 *
+	 * --type=<value>
+	 * : Set the type for this source. This must be the internal name of the used service, e.g., local or ftp.
+	 *
+	 * --fields=<value>
+	 * : Fields to configure this source as JSON.
+	 *
+	 * @param array<int,string>    $names Array with names, which might be given as parameter on CLI-command.
+	 * @param array<string,string> $arguments List of parameter to use for the given URLs.
+	 *
+	 * @return void
+	 */
+	public function add_external_source( array $names = array(), array $arguments = array() ): void {
+		// get the fields.
+		$fields = json_decode( $arguments['fields'], true );
+
+		// bail if fields is not an array.
+		if ( ! is_array( $fields ) ) {
+			\WP_CLI::error( __( 'Wrong fields value given!', 'external-files-in-media-library' ) );
+		}
+
+		// add the external source.
+		$term_id = Taxonomy::get_instance()->add( $arguments['type'], $names[0], $fields );
+
+		// bail if no entry could be saved.
+		if ( 0 === $term_id ) {
+			\WP_CLI::error( __( 'Wrong fields value given!', 'external-files-in-media-library' ) );
+		}
+
+		// show ok message.
+		\WP_CLI::success( __( 'The external source has been added.', 'external-files-in-media-library' ) );
+	}
+
+	/**
+	 * Delete external sources by their names.
+	 *
+	 * <Names>
+	 * : List of names of external sources to delete.
+	 *
+	 * @param array<int,string> $names List of names.
+	 *
+	 * @return void
+	 */
+	public function delete_external_source( array $names = array() ): void {
+		// get the taxonomy object.
+		$taxonomy_obj = Taxonomy::get_instance();
+
+		// check each given name.
+		foreach ( $names as $name ) {
+			// get the term by the given name.
+			$query  = array(
+				'taxonomy'   => $taxonomy_obj->get_name(),
+				'hide_empty' => false,
+				'meta_query' => array(
+					array(
+						'key'     => 'path',
+						'value'   => $name,
+						'compare' => '=',
+					),
+				),
+			);
+			$result = new WP_Term_Query( $query );
+
+			// bail on no results.
+			if ( empty( $result->terms ) ) {
+				continue;
+			}
+
+			// delete the entry.
+			$taxonomy_obj->delete( $result->terms[0]->term_id );
+		}
+
+		// show ok message.
+		\WP_CLI::success( __( 'Given external sources has been deleted.', 'external-files-in-media-library' ) );
+	}
+
+	/**
+	 * Run sync for given external sources.
+	 *
+	 * <Names>
+	 * : Names of external sources to sync.
+	 *
+	 * @param array<int,string> $names List of names.
+	 *
+	 * @return void
+	 */
+	public function sync( array $names = array() ): void {
+		// bail if sync is enabled.
+		if ( 0 === absint( get_option( 'eml_sync' ) ) ) {
+			// show hint.
+			\WP_CLI::success( __( 'Synchronisation is not enabled in the plugin settings.', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// bail if no names are given.
+		if ( empty( $names ) ) {
+			// show hint.
+			\WP_CLI::error( __( 'No external sources given!', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// get the taxonomy object.
+		$taxonomy_obj = Taxonomy::get_instance();
+
+		// get the synchronisation object.
+		$sync_obj = Synchronization::get_instance();
+
+		// sync each given name.
+		foreach ( $names as $name ) {
+			// get the term by the given name.
+			$query  = array(
+				'taxonomy'   => $taxonomy_obj->get_name(),
+				'hide_empty' => false,
+				'meta_query' => array(
+					array(
+						'key'     => 'path',
+						'value'   => $name,
+						'compare' => '=',
+					),
+				),
+			);
+			$result = new WP_Term_Query( $query );
+
+			// bail on no results.
+			if ( empty( $result->terms ) ) {
+				continue;
+			}
+
+			// get the term ID.
+			$term_id = $result->terms[0]->term_id;
+
+			// get the term data.
+			$term_data = $taxonomy_obj->get_entry( $term_id );
+
+			// run the synchronisation for this external source.
+			$sync_obj->sync( $name, $term_data, $term_id );
+		}
+
+		// show ok message.
+		\WP_CLI::success( __( 'Synchronisation for given external sources have been run.', 'external-files-in-media-library' ) );
+	}
+
+	/**
+	 * Delete synced files for given external sources.
+	 *
+	 * <Names>
+	 * : Names of external sources to sync.
+	 *
+	 * @param array<int,string> $names List of names.
+	 *
+	 * @return void
+	 */
+	public function delete_synced_files( array $names = array() ): void {
+		// bail if sync is enabled.
+		if ( 0 === absint( get_option( 'eml_sync' ) ) ) {
+			// show hint.
+			\WP_CLI::success( __( 'Synchronisation is not enabled in the plugin settings.', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// bail if no names are given.
+		if ( empty( $names ) ) {
+			// show hint.
+			\WP_CLI::error( __( 'No external sources given!', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// get the taxonomy object.
+		$taxonomy_obj = Taxonomy::get_instance();
+
+		// get the synchronisation object.
+		$sync_obj = Synchronization::get_instance();
+
+		// delete files for each given name.
+		foreach ( $names as $name ) {
+			// get the term by the given name.
+			$query  = array(
+				'taxonomy'   => $taxonomy_obj->get_name(),
+				'hide_empty' => false,
+				'meta_query' => array(
+					array(
+						'key'     => 'path',
+						'value'   => $name,
+						'compare' => '=',
+					),
+				),
+			);
+			$result = new WP_Term_Query( $query );
+
+			// bail on no results.
+			if ( empty( $result->terms ) ) {
+				continue;
+			}
+
+			// get the term ID.
+			$term_id = $result->terms[0]->term_id;
+
+			// delete the synced files of this external source.
+			$sync_obj->delete_synced_files( $term_id, $taxonomy_obj->get_name() );
+		}
+
+		// show ok message.
+		\WP_CLI::success( __( 'Synchronised files have been deleted.', 'external-files-in-media-library' ) );
+	}
+
+	/**
+	 * Change the export state for given external sources.
+	 *
+	 * <Names>
+	 * : Names of external sources.
+	 *
+	 * [--enable]
+	 * : Enable the export for the given external sources.
+	 *
+	 * [--disable]
+	 * : Disable the export for the given external sources.
+	 *
+	 * Hint: enabling the export to an external source can only be successful
+	 * if it has already been configured for export.
+	 *
+	 * @param array<int,string>    $names List of names.
+	 * @param array<string,string> $arguments The arguments.
+	 *
+	 * @return void
+	 */
+	public function change_export_state( array $names = array(), array $arguments = array() ): void {
+		// bail if sync is enabled.
+		if ( 0 === absint( get_option( 'eml_export' ) ) ) {
+			// show hint.
+			\WP_CLI::success( __( 'Export is not enabled in the plugin settings.', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// bail if no names are given.
+		if ( empty( $names ) ) {
+			// show hint.
+			\WP_CLI::error( __( 'No external sources given!', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// bail if no arguments are set.
+		if ( empty( $arguments ) ) {
+			// show hint.
+			\WP_CLI::error( __( 'No setting is given given!', 'external-files-in-media-library' ) );
+
+			// do nothing more.
+			return;
+		}
+
+		// get the taxonomy object.
+		$taxonomy_obj = Taxonomy::get_instance();
+
+		// get the export object.
+		$export_obj = Export::get_instance();
+
+		// enable export for each given name.
+		foreach ( $names as $name ) {
+			// get the term by the given name.
+			$query  = array(
+				'taxonomy'   => $taxonomy_obj->get_name(),
+				'hide_empty' => false,
+				'meta_query' => array(
+					array(
+						'key'     => 'path',
+						'value'   => $name,
+						'compare' => '=',
+					),
+				),
+			);
+			$result = new WP_Term_Query( $query );
+
+			// bail on no results.
+			if ( empty( $result->terms ) ) {
+				continue;
+			}
+
+			// get the term ID.
+			$term_id = $result->terms[0]->term_id;
+
+			// enable the export to this external source.
+			$export_obj->set_state_for_term( $term_id, isset( $arguments['enable'] ) ? 1 : 0 );
+		}
+
+		// show ok message.
+		\WP_CLI::success( __( 'Settings for export to external sources has been changed.', 'external-files-in-media-library' ) );
 	}
 }